python/requirements.txt
cryptography
ipcalc
netifaces
ping3
six
systemd-python
validators
Jinja2
PyYAML
pip
toml
wheel
py
tornado
platformdirs
setuptools
packaging
MarkupSafe
pluggy
pyparsing
pycparser
decorator
requests
certifi
urllib3
idna
snowflake
